pyproject.toml
[project]
name = "handwriting-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-htmx>=1.21.0",
    "django>=5.1.4",
    "huggingface-hub>=0.27.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
